requirements.txt
tensorflow
pillow
flask
numpy
